Added conn.destroy to fix ECONNRESET bug
Found a bug where ECONNRESET would be thrown after the socket on port 25 abruptely closes. Adding conn.destroy() after conn.removeAllListeners() will ensure the port is closed then and there, preventing an error to be thrown when it eventually closes. Minor testing seems to indicate that .destroy() is not needed on the 'error' and 'false' events, but I could be mistaken.